Fix build on recent Mac OS X.
Recent versions of the OS X SDK remove some of the old Quartz interfaces,
so that SDL no longer builds when using MACOSX_DEPLOYMENT_TARGET=10.4.
Since 10.4 is ancient, bump this up to 10.7 as the oldest supported
version (2011 is a long time ago now), and apply a patch during build
for SDL bug 2085.

Thanks to Linguica for the bug report. This fixes
chocolate-doom/chocolate-doom#584.